[feat#17] border color 수정 및 수입 수정 기능 추가
[feat#17] border color 수정 및 수입 수정 기능 추가
