+++ src/components/Header.jsx
@@ -10,10 +10,10 @@ const Header = ({ title, subtitle }) => {
         variant="h2"
         color={colors.grey[100]}
         fontWeight="bold"
-        sx={{ m: "0 0 5px 0" }}
-      >
+        sx={{ m: "0 0 5px 0" }}>
         {title}
       </Typography>
+      
       <Typography variant="h5" color={colors.greenAccent[400]}>
         {subtitle}
       </Typography>
