+++ src/components/IntroSwiper.jsx
@@ -9,7 +9,7 @@ import Sub_one from '../assets/images/side_one.svg';
 import Sub_two from '../assets/images/side_two.svg';
 import Sub_three from '../assets/images/side_three.svg';
 import Sub_four from '../assets/images/side_four.svg';
-import "./IntroSwiper.css";
+import "../components/css/IntroSwiper.css"
 
 SwiperCore.use([Autoplay,Pagination]);
 
@@ -29,7 +29,7 @@ function IntroSwiper(props){
 
       <SwiperSlide>
           <div className = "intro">
-               <img className = "main-image"src={Sub_one} />
+               <img className = "main-image"src={Sub_one} alt="one"/>
                <p className = "main-one">#잃어버린 #텅장을 #찾아서</p>
                <p className = "main-two"> SHOW ME THE MONEY </p>
                <p className = "main-three">줄줄새는 내돈<br/>꽉 잡아줄 나만의 가계부</p>
@@ -39,7 +39,7 @@ function IntroSwiper(props){
                
       <SwiperSlide>
           <div className = "intro">
-               <img className = "main-image"src={Sub_two} />
+               <img className = "main-image"src={Sub_two} alt="two"/>
                <p className = "main-one">#잃어버린 #텅장을 #찾아서</p>
                <p className = "main-two"> SHOW ME THE MONEY </p>
                <p className = "main-three">줄줄새는 내돈<br/>꽉 잡아줄 나만의 가계부</p>
@@ -48,7 +48,7 @@ function IntroSwiper(props){
       </SwiperSlide>
       <SwiperSlide>
           <div className = "intro">
-               <img className = "main-image"src={Sub_three} />
+               <img className = "main-image"src={Sub_three} alt="three"/>
                <p className = "main-one">#잃어버린 #텅장을 #찾아서</p>
                <p className = "main-two"> SHOW ME THE MONEY </p>
                <p className = "main-three">줄줄새는 내돈<br/>꽉 잡아줄 나만의 가계부</p>
@@ -57,7 +57,7 @@ function IntroSwiper(props){
       </SwiperSlide>
       <SwiperSlide>
           <div className = "intro">
-               <img className = "main-image"src={Sub_four} />
+               <img className = "main-image"src={Sub_four} alt="four"/>
                <p className = "main-one">#잃어버린 #텅장을 #찾아서</p>
                <p className = "main-two"> SHOW ME THE MONEY </p>
                <p className = "main-three">줄줄새는 내돈<br/>꽉 잡아줄 나만의 가계부</p>
