+++ src/pages/challenge/index.jsx
@@ -1,16 +1,17 @@
+import { useState, useRef, useEffect} from "react";
 import { Box } from "@mui/material";
 import Header from "../../components/Header";
-import './index.css'
 import Button from 'react-bootstrap/Button';
-import Ch from "../../assets/images/sm.svg"
-import { useState, useRef, useEffect} from "react";
 import Sidebar from '../global/Sidebar';
 import Topbar from '../global/Topbar';
+
+import Ch from "../../assets/images/sm.svg"
 import OutlinedInput from '@mui/material/OutlinedInput';
 import InputLabel from '@mui/material/InputLabel';
 import InputAdornment from '@mui/material/InputAdornment';
 import FormControl from '@mui/material/FormControl';
 import axios from "axios";
+import './index.css'
 
 const Challenge = () => {
 
@@ -43,7 +44,6 @@ const Challenge = () => {
     const newdata = {...data}
     newdata[e.target.id] = e.target.value
     setData(newdata)
-    console.log(newdata)
   }
 
    
@@ -73,7 +73,6 @@ const Challenge = () => {
         <h2 className="sub-three">금액을 설정하여, 같이 모아봐요!</h2>
         <p className="sub-four">STEP 1.  금액을 설정해주세요.<br/>
           STEP 2.  현재 지출한 금액을 계산하여, 얼마나 남았는지 알려드릴게요.</p>
-        {/* <input type="number" className="input-value"></input> */}
 
       <Box component="form"
       sx={{
@@ -81,12 +80,10 @@ const Challenge = () => {
         "& .MuiOutlinedInput":{"& > fieldset": {border: '1px solid green'}}
       }}
       noValidate
-      autoComplete="off"
-     
-      >
+      autoComplete="off">
+
       <FormControl 
-        fullWidth sx={{ mt: 12}}
-        >
+        fullWidth sx={{ mt: 12}}>
           <InputLabel htmlFor="outlined-adornment-amount">금액</InputLabel>
           <OutlinedInput
             id="budget"
@@ -102,9 +99,7 @@ const Challenge = () => {
         </FormControl>
       </Box>
         <Button type="submit" className="input-button" 
-        onClick={() => {
-          totalsubmit();
-        }}> 확인 </Button>
+        onClick={() => { totalsubmit(); window.location.reload(); }}> 확인 </Button>
         <div className="budget">
             <span>목표금액 : {budget}</span>
             <p >지출금액 : {list}</p>
