+++ src/pages/dashboard/index.jsx
@@ -1,8 +1,7 @@
 import React, {useState,useEffect} from 'react';
 import axios from 'axios';
-import { Box, IconButton, Typography, useTheme } from "@mui/material";
+import { Box, Typography, useTheme } from "@mui/material";
 import { tokens } from "../../theme";
-import DownloadOutlinedIcon from "@mui/icons-material/DownloadOutlined";
 import AttachMoneyIcon from '@mui/icons-material/AttachMoney';
 import MoneyOffIcon from '@mui/icons-material/MoneyOff';
 import Header from "../../components/Header";
@@ -13,6 +12,7 @@ import Sidebar from '../global/Sidebar';
 import Topbar from '../global/Topbar';
 
 const Dashboard = () => {
+
   const theme = useTheme();
   const colors = tokens(theme.palette.mode);
   const [isSidebar, setIsSidebar] = useState(true);
@@ -67,7 +67,7 @@ const Dashboard = () => {
     <div className="app">
           <Sidebar isSidebar={isSidebar} />
     <main className="content">
-          <Topbar setIsSidebar={setIsSidebar}  />
+          <Topbar setIsSidebar={setIsSidebar} />
     </main>
  
     <div className="size">
@@ -82,16 +82,14 @@ const Dashboard = () => {
         display="grid"
         gridTemplateColumns="repeat(12, 1fr)"
         gridAutoRows="140px"
-        gap="20px"
-      >
+        gap="20px">
         {/* ROW 1 */}
         <Box
           gridColumn="span 3"
           backgroundColor={colors.primary[400]}
           display="flex"
           alignItems="center"
-          justifyContent="center"
-        >
+          justifyContent="center">
           <StatBox
             title= {incomedata.total_income+"원"}
             subtitle="수입"
@@ -108,8 +106,7 @@ const Dashboard = () => {
           backgroundColor={colors.primary[400]}
           display="flex"
           alignItems="center"
-          justifyContent="center"
-        >
+          justifyContent="center">
           
           <StatBox
             title = {data.total_spending+"원"}
@@ -117,7 +114,7 @@ const Dashboard = () => {
             progress="0.50"
             icon={
               <MoneyOffIcon
-                sx={{ color: colors.greenAccent[600], fontSize: "26px" }}
+                sx={{ color: colors.greenAccent[600], fontSize: "26px" }} 
               />
             }
           />
@@ -145,8 +142,7 @@ const Dashboard = () => {
           backgroundColor={colors.primary[400]}
           display="flex"
           alignItems="center"
-          justifyContent="center"
-        >
+          justifyContent="center">
           <StatBox
             title={threespenddata.total_three_month_ago_spending+"원"}
             subtitle="3개월 전 지출"
@@ -163,47 +159,43 @@ const Dashboard = () => {
         <Box
           gridColumn="span 8"
           gridRow="span 2"
-          backgroundColor={colors.primary[400]}
-        >
+          backgroundColor={colors.primary[400]}>
           <Box
             mt="25px"
             p="0 30px"
             display="flex "
             justifyContent="space-between"
-            alignItems="center"
-          >
+            alignItems="center">
 
           </Box>
           <Box height="280px" m="-37px 0 0 70px">
             <BarChart isDashboard={true} />
           </Box>
         </Box>
-       
 
         {/* ROW 3 */}
         <Box
           gridColumn="span 4"
           gridRow="span 2"
           backgroundColor={colors.primary[400]}
-          p="30px"
-        >
+          p="30px">
           <Typography variant="h5" fontWeight="600">
             한달 전보다 얼마나 지출이 적을까?
           </Typography>
+
           <Box
             display="flex"
             flexDirection="column"
             alignItems="center"
-            mt="25px"
-          >
+            mt="25px">
             <ProgressCircle size="125" />
             <Typography
               variant="h5"
               color={colors.greenAccent[500]}
-              sx={{ mt: "15px" }}
-            >
+              sx={{ mt: "15px" }}>
               전달보다 <b>{totaldata.comparison_total_spending}</b>원 덜 소비 하셨어요!
             </Typography>
+            
             <Typography>소비를 덜 하고싶다면 챌린지 참여해보세요 </Typography>
           </Box>
         </Box>
